Redone sum equation and parseFloat for decimals
I added parseFloat to several area to keep the terms as numbers instead of strings so that deciamls would be calculated correctly. I orginally just did it once on numberButtons but then '.' did not work. So i decided that parseFloat 3 more times in the code was better code practice than add an if statement for just '.'. Perhaps The next best option was to change the id for '.' and make a seperate eventlistener. This would cause more work now (and probably some debugging adding strings to numbers?) but might save me work for when I want to limit each term to only once possible use of the decimal place